ChainRecursive.jl

Documentation for ChainRecursive.jl

ChainRecursive.chain

ChainRecursive.@chain

ChainRecursive.chain — Method.

@chain(e)

Will chain all eligible block in your code, recursively.

Within each block, lines are reduced with link. link has two behaviors. It will call bare functions on the result of the previous line.

julia> using ChainRecursive

julia> @chain begin
           1
           vcat
        end
1-element Array{Int64,1}:
 1

It will reinterpret _ within expressions as the result of the previous line.

julia> using ChainRecursive

julia> @chain begin
           1
           vcat(2, _)
       end
2-element Array{Int64,1}:
 2
 1

It will recur.

julia> using ChainRecursive

julia> @chain begin
           1
           if _ == 1
               _ + 1
               vcat(_, 3)
           end
       end
2-element Array{Int64,1}:
 2
 3

Cannot chain blocks with assignments, bindings, or updates. They will be skipped. Once you give a name to something, use it! If your chains get too long, I highly recommend breaking them up with descriptive assignments.

julia> using ChainRecursive

julia> @chain begin
           a = 1
           _
       end
ERROR: UndefVarError: _ not defined

julia> @chain begin
            a += 1
            _
       end
ERROR: UndefVarError: _ not defined

julia> @chain begin
           function test_function(x) 1 end
           _
       end
ERROR: UndefVarError: _ not defined

julia> @chain begin
           type cant_chain_types end
           _
       end
ERROR: UndefVarError: _ not defined

However, assignments, bindings, and updates will not prevent recursion.

julia> using ChainRecursive

julia> @chain begin
           a = 1
           begin
               a + 1
               vcat(_, 3)
           end
       end
2-element Array{Int64,1}:
 2
 3

In addition, type definitions will not be recurred into.

julia> using ChainRecursive

julia> @chain type test_type
           a
           b::_
       end
ERROR: UndefVarError: _ not defined

source

ChainRecursive.@chain — Macro.

See documentation of chain

source